requirements.txt
#==This
#==platform:
absl-py=2.1.0=pypi_0==
blas=1.0=openblas==
bottleneck=1.3.7=py312ha86b861_0==
brotli=1.0.9=h80987f9_8==
brotli-bin=1.0.9=h80987f9_8==
brotli-python=1.0.9=py312h313beb8_8==
bzip2=1.0.8=h80987f9_6==
ca-certificates=2024.9.24=hca03da5_0==
certifi=2024.8.30=py312hca03da5_0==
cffi=1.17.1=pypi_0==
charset-normalizer=3.3.2=pyhd3eb1b0_0==
contourpy=1.2.0=py312h48ca7d4_0==
cpuonly=2.0=0==
cycler=0.11.0=pyhd3eb1b0_0==
cython=3.0.11=pypi_0==
etils=1.9.4=pypi_0==
expat=2.6.3=h313beb8_0==
fasteners=0.19=pypi_0==
filelock=3.13.1=py312hca03da5_0==
fonttools=4.51.0=py312h80987f9_0==
freetype=2.12.1=h1192e45_0==
fsspec=2024.9.0=pypi_0==
glfw=2.7.0=pypi_0==
idna=3.7=py312hca03da5_0==
imageio=2.35.1=pypi_0==
importlib-resources=6.4.5=pypi_0==
jinja2=3.1.4=py312hca03da5_0==
jpeg=9e=h80987f9_3==
kiwisolver=1.4.4=py312h313beb8_0==
lcms2=2.12=hba8e193_0==
lerc=3.0=hc377ac9_0==
libbrotlicommon=1.0.9=h80987f9_8==
libbrotlidec=1.0.9=h80987f9_8==
libbrotlienc=1.0.9=h80987f9_8==
libcxx=14.0.6=h848a8c0_0==
libdeflate=1.17=h80987f9_1==
libffi=3.4.4=hca03da5_1==
libgfortran=5.0.0=11_3_0_hca03da5_28==
libgfortran5=11.3.0=h009349e_28==
libjpeg-turbo=2.0.0=h1a28f6b_0==
libopenblas=0.3.21=h269037a_0==
libpng=1.6.39=h80987f9_0==
libtiff=4.5.1=h313beb8_0==
libwebp-base=1.3.2=h80987f9_0==
llvm-openmp=14.0.6=hc6e5704_0==
lz4-c=1.9.4=h313beb8_1==
markupsafe=2.1.3=py312h80987f9_0==
matplotlib=3.9.2=py312hca03da5_0==
matplotlib-base=3.9.2=py312h2df2da3_0==
mpmath=1.3.0=py312hca03da5_0==
mujoco=3.2.3=pypi_0==
mujoco-py=2.1.2.14=pypi_0==
ncurses=6.4=h313beb8_0==
networkx=3.3=py312hca03da5_0==
numexpr=2.8.7=py312h0f3ea24_0==
numpy=1.26.4=py312h7f4fdc5_0==
numpy-base=1.26.4=py312he047099_0==
openjpeg=2.5.2=h54b8e55_0==
openssl=3.0.15=h80987f9_0==
packaging=24.1=py312hca03da5_0==
pandas=2.2.2=py312hd77ebd4_0==
pillow=10.4.0=py312h80987f9_0==
pip=24.2=py312hca03da5_0==
pycparser=2.22=pypi_0==
pyopengl=3.1.7=pypi_0==
pyparsing=3.1.2=py312hca03da5_0==
pysocks=1.7.1=py312hca03da5_0==
python=3.12.4=h99e199e_1==
python-dateutil=2.9.0post0=py312hca03da5_2==
python-tzdata=2023.3=pyhd3eb1b0_0==
pytorch=2.4.1=py3.12_0==
pytorch-mutex=1.0=cpu==
pytz=2024.1=py312hca03da5_0==
pyyaml=6.0.2=py312h80987f9_0==
readline=8.2=h1a28f6b_0==
requests=2.32.3=py312hca03da5_0==
setuptools=75.1.0=py312hca03da5_0==
six=1.16.0=pyhd3eb1b0_1==
sqlite=3.45.3=h80987f9_0==
sympy=1.13.2=py312hca03da5_0==
tk=8.6.14=h6ba3021_0==
torchaudio=2.4.1=py312_cpu==
torchvision=0.19.1=py312_cpu==
tornado=6.4.1=py312h80987f9_0==
typing_extensions=4.11.0=py312hca03da5_0==
tzdata=2024b=h04d1e81_0==
unicodedata2=15.1.0=py312h80987f9_0==
urllib3=2.2.3=py312hca03da5_0==
wheel=0.44.0=py312hca03da5_0==
xz=5.4.6=h80987f9_1==
yaml=0.2.5=h1a28f6b_0==
zipp=3.20.2=pypi_0==
zlib=1.2.13=h18a0788_1==
zstd=1.5.6=hfb09047_0==
